Add AJAX support to form submission
Submitting values to Thingspeak now works over AJAX, returning you to your entry form, and giving
a confirmation popup for 2 seconds.

Due to the way cross-domain AJAX requests work (or, rather, don't), it's not possible at this time
to parse the response to the post to see if it was successful or not.  So we're just going to assume it was.